requirements.txt
Django==5.0.6
taggit==4.1.0
django-ckeditor==6.7.0
paypal==1.2.5
django-paypal==2.0
gunicorn==20.1.0
whitenoise==6.6.0
python-decouple==3.8
psycopg2-binary==2.9.7
Pillow==10.0.0
django-jazzmin==2.6.0
